infernet: Inferential Models for Many Different Types of Networks

A set of tools for testing networks. It includes functions for univariate and multivariate conditional uniform graph and quadratic assignment procedure testing, and multiple regression quadratic assignment procedure (MRQAP) models for a range of families, with random and fixed effects, missing data, and cognitive social structures. Built on the 'manynet' package, all functions operate with matrices, edge lists, and 'igraph', 'network', and 'tidygraph' objects, and on one-mode and two-mode (bipartite) networks.

See also

Author

Maintainer: James Hollway james.hollway@graduateinstitute.ch (ORCID) (IHEID) [contributor]

Authors:

James Hollway james.hollway@graduateinstitute.ch (ORCID) (IHEID) [contributor]

Other contributors:

Robert Krause (ORCID) [contributor]