Tests of network measures

These functions conduct tests of any network-level statistic:

test_random() performs a conditional uniform graph (CUG) test of a measure against a distribution of measures on random networks of the same dimensions.

test_permutation() performs a quadratic assignment procedure (QAP) test of a measure against a distribution of measures on permutations of the original network.

test_random(
  .data,
  FUN,
  ...,
  times = 1000,
  strategy = "sequential",
  verbose = FALSE
)

test_configuration(
  .data,
  FUN,
  ...,
  times = 1000,
  strategy = "sequential",
  verbose = FALSE
)

test_permutation(
  .data,
  FUN,
  ...,
  times = 1000,
  strategy = "sequential",
  verbose = FALSE
)

Arguments

.data

A manynet-consistent network (see manynet::as_tidygraph()), or a list of such networks. When a list is supplied the model is fit jointly; graphs missing any predictor are dropped with a warning.

FUN

A graph-level statistic function to test.

...

Additional arguments to be passed on to FUN, e.g. the name of the attribute.

times

Integer. Number of permutations for the null distribution. 1000 is the default; publication-ready work usually needs 1000-10000.

See also

Examples

marvel_friends <- to_unsigned(ison_marvel_relationships)
#> Error in to_unsigned(ison_marvel_relationships): could not find function "to_unsigned"
marvel_friends <- to_giant(marvel_friends) %>% 
  to_subgraph(PowerOrigin == "Human")
#> Error in to_giant(marvel_friends) %>% to_subgraph(PowerOrigin == "Human"): could not find function "%>%"
(cugtest <- test_random(marvel_friends, manynet::net_heterophily, attribute = "Attractive",
   times = 200))
#> ℹ `net_heterophily()` was moved since manynet 2.0.0 to netrics.
#> Install netrics with: `install.packages('netrics')`
#> Error: Please update your code to use the new function.
# plot(cugtest)
# (qaptest <- test_permutation(marvel_friends, 
#                 manynet::net_heterophily, attribute = "Attractive",
#                 times = 200))
# plot(qaptest)